pyproject.toml
[tool.poetry]
name = "aikeyboard"
version = "0.1.0"
description = ""
authors = ["PranitSh <ppshah2023@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8, <=3.10.6"
pyautogui = "^0.9.53"
pytesseract = "^0.3.10"
pillow = "^9.5.0"
opencv-python = "^4.7.0.72"
rtree = "^1.0.1"
pynput = "^1.7.6"
google-cloud-core = "^2.3.2"


[tool.poetry.group.dev.dependencies]
pyinstaller = "^5.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
